src/requirements.txt
Django==3.0.1
djangorestframework==3.11.0
jsonschema==3.2.0